% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEWAquarius.R
\name{AQWPLoad}
\alias{AQWPLoad}
\title{Function to load in an Aquarius json file, downloaded from https://water.data.sa.gov.au, possibly using AWQPDownload()}
\usage{
AQWPLoad(filename, qual_codes = TRUE, long_format = TRUE)
}
\arguments{
\item{filename}{A file downloaded from the Export Data tab on https://water.data.sa.gov.au, or using AQWPDownload()}

\item{qual_codes}{TRUE/FALSE to return quality codes. Defaults to true}

\item{long_format}{TRUE/FALSE to return data in long format, rather than wide (e.g. a spreadsheet). Long is useful for plotting with ggplot}
}
\value{
A tibble with the data in the file
}
\description{
Function to load in an Aquarius json file, downloaded from https://water.data.sa.gov.au, possibly using AWQPDownload()
}
\examples{
\dontrun{
AQWPLoad("AQWP.json")
}

}
