% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydstra.R
\name{HydstraExport}
\alias{HydstraExport}
\title{Function to export Hydstra data, similar to HYCSV}
\usage{
HydstraExport(sites, varFrom, varTo, startdate = "09:00_01/01/1980",
  enddate = "09:00_01/01/1980", Interval = "DAY", path = NULL,
  file = "Auto.ini")
}
\arguments{
\item{sites}{A station number or vector of station numbers, as a string (e.g. "A4261001")}

\item{startdate}{First period of data, including the time, e.g. "09:00_01/01/2017".}

\item{enddate}{Last period of data, including the time, e.g. "09:00_30/07/2017".}

\item{Interval}{Period of data required, e.g. HOUR, DAY, MONTH}

\item{path}{Where to save the output time series, defaults to getwd()}

\item{VarFrom}{A variable that exists for the station to convert from, for example water level, as a string e.g. "100.00".}

\item{VarTo}{A variable can be output for the station, for example flow, as a string e.g. "141.00".}
}
\value{
A file is created with the data requested. Nothing is returned to the R environment
}
\description{
Function to export Hydstra data, similar to HYCSV
}
\examples{
HydstraExport("A2390519","100.00","141.00")

}
