% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILOWriteforSource.R
\name{read_res.csv}
\alias{read_res.csv}
\title{Function to Read a Source .res.csv File
Returns data (as a Data Frame,  Zoo, or tibble) as a time Series with all Results
Read Source .res.csv file into a data table or zoo time series}
\usage{
read_res.csv(resFile, returnType = "df")
}
\arguments{
\item{resFile}{A character string representing the full file path of the .res.csv file}

\item{returnType}{A character string to set the return type: "z", "t", "df". If not matching "t" (tibble) or "z" (zoo), data frame returned.}
}
\value{
Data in the format selected with all data read in from the Source .res.csv file
}
\description{
Function to Read a Source .res.csv File
Returns data (as a Data Frame,  Zoo, or tibble) as a time Series with all Results
Read Source .res.csv file into a data table or zoo time series
}
\examples{
X = read_res.csv("SWTools/extdata/Scenario1.res.csv",returnType="t")

}
