% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{SILODownload}
\alias{SILODownload}
\title{Download SILO data}
\usage{
SILODownload(SiteList, username, password, path = getwd(),
  startdate = "18890101", enddate = NULL)
}
\arguments{
\item{SiteList}{A station number or vector of station numbers, as a string (e.g. "24001")}

\item{username}{SILO user name}

\item{password}{SILO password}

\item{path}{Where to save the output. Will default to getwd() if not specified}

\item{startdate}{First day of data, in the format "YYYYMMDD". Will default to the first day of the record "18890101" if not specified}

\item{enddate}{Last day of data, in the format "YYYYMMDD". Will default to yesterday if not specified}
}
\value{
A file for each station will be saved to path, named station number.txt. Nothing is returned to the R environment.
}
\description{
Download SILO data
}
\examples{
SILODownload("24001")
SILODownload("24001","C:/SILO/","20170701","20170801")


}
