% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{SILOMap}
\alias{SILOMap}
\title{Plot a map of the SILO station locations}
\usage{
SILOMap(SILO, filename = NULL)
}
\arguments{
\item{SILO}{a list of sites with SILO data, as created by SILOLoad()}

\item{filename}{optional, filename to write the plot to, including extension. Filename can include full path or sub folders.}
}
\value{
a google map of the SILO station locations
}
\description{
Plot a map of the SILO station locations
}
\examples{
\dontrun{
X<-SILOLoad(c("24001","24002","24003"),path="./SWTools/extdata")
p<-SILOMap(X)
}

}
