% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{SILOReport}
\alias{SILOReport}
\title{Write SILO data report to word document}
\usage{
SILOReport(SILO, filename)
}
\arguments{
\item{SILO}{a list of sites with SILO data, as created by SILOLoad()}

\item{filename}{filename to write the report to.}
}
\description{
Write SILO data report to word document
}
\examples{
X<-LoadSILO(c("24001","24002","24003")
SILOReport(X,"C:/Output/MyReport.docx")
}
