% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{SILOSiteSummary}
\alias{SILOSiteSummary}
\title{Produce a table summarising  SILO sites}
\usage{
SILOSiteSummary(SILO)
}
\arguments{
\item{SILO}{a list of sites with SILO data, as created by SILOLoad()}
}
\value{
a dataframe with the following columns
Site - site name
Station - station number
StartDate - date of the first good quality rainfall data
EndDate - date of the last good quality rainfall data
PctMissing - percentage of days that do not have good quality code between StartDate and EndDate
AnnualRainfall - Mean annual rainfall in mm
Latitute - Latitude
Longitude - Longitude
}
\description{
Produce a table summarising  SILO sites
}
\examples{
X<-SILOLoad(c("24001","24002","24003"))
d<-SILOSummary(X)

}
