% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILOThiessenShp.R
\name{SILOThiessenShp}
\alias{SILOThiessenShp}
\title{Function to generate Thiessen polygons from SILO sites}
\usage{
SILOThiessenShp(SILOdata, path, shpname)
}
\arguments{
\item{SILOdata}{- data loaded from SILO based on site list}

\item{path}{- file path to save Thiessen polygon shapefile}

\item{shpname}{- filename to save ESRI shapefile (no extension)}
}
\value{
nothing to the environment. Shape file saved to path \\ shpname
}
\description{
Function to generate Thiessen polygons from SILO sites
}
\examples{
X<-SILOLoad(c("24001","24002","24003"),path="SWTools/extdata")
SILOThiessenShp(X,".","Theissens")
}
