% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerGetInputSets}
\alias{VeneerGetInputSets}
\title{Get vector of InputSets}
\usage{
VeneerGetInputSets(baseURL = "http://localhost:9876")
}
\arguments{
\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
vector containing info on Input Sets in the model
}
\description{
Get vector of InputSets
}
\examples{
\dontrun{
VeneerGetInputSets()
}

}
