% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerGetNodesbyType}
\alias{VeneerGetNodesbyType}
\title{Get a vector of node names for a given type}
\usage{
VeneerGetNodesbyType(NodeType, baseURL = "http://localhost:9876")
}
\arguments{
\item{NodeType}{The node to return the names of. The icon in /network is searched for this name}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
vector of node names matching the specified node type
}
\description{
Get a vector of node names for a given type
}
\examples{
\dontrun{
VeneerGetNodesbyType("Weir")
}

}
