% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerGetPiecewise}
\alias{VeneerGetPiecewise}
\title{Get data from a Source piecewise table using Veneer}
\usage{
VeneerGetPiecewise(pw_table, baseURL = "http://localhost:9876")
}
\arguments{
\item{pw_table}{The name of the piecewise linear variable, without the $}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
a matrix with the data from the piecewise table.

some text to hopefully trick
roxygen into working
}
\description{
Get data from a Source piecewise table using Veneer
}
\examples{
\dontrun{
VeneerGetPiecewise(data,"pw_table")
}
 
 
}
