% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerGetTS}
\alias{VeneerGetTS}
\title{Get a time series result from Source using Veneer}
\usage{
VeneerGetTS(TSURL, baseURL = "http://localhost:9876")
}
\arguments{
\item{TSURL, }{the URL of the time series to retrieve}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
a zoo time series of the data

The URL of the time series must be specified, by interrogation using a browser or other analysis. 
By default Source returns SI units. Some conversion is undertaken:
* Flow converted to ML/d
* Volume converted to ML
* Area converted to ha

Spaces are OK, like in the example below (dont need to insert %20 for example).
}
\description{
Get a time series result from Source using Veneer
}
\examples{
\dontrun{
VeneerGetTS("/runs/latest/location/EndofSystem/element/Downstream Flow/variable/Flow")
}

}
