% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerGetTSbyNode}
\alias{VeneerGetTSbyNode}
\title{Get all time series recorded in Source for a given node}
\usage{
VeneerGetTSbyNode(Node, run = "latest", baseURL = "http://localhost:9876")
}
\arguments{
\item{Node}{Name of node to retrieve Time Series for}

\item{run}{Which run to retrieve from. Defaults to the latest}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
a zoo time series, with each variable as a column
}
\description{
Get all time series recorded in Source for a given node
}
\examples{
\dontrun{
VeneerGetTSbyNode("Storage 1")
}

}
