% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerGetTSbyVariable}
\alias{VeneerGetTSbyVariable}
\title{Get all time series recorded in Source of a given variable type}
\usage{
VeneerGetTSbyVariable(
  variable = "Flow",
  run = "latest",
  baseURL = "http://localhost:9876"
)
}
\arguments{
\item{variable}{Which variable to retrieve. Defaults to Flow.}

\item{run}{Which run to retrieve from. Defaults to the latest}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
a zoo time series, with each output as a column
}
\description{
Get all time series recorded in Source of a given variable type
}
\examples{
\dontrun{
VeneerGetTSbyVariable() #returns all flow outputs recorded in the latest run
VeneerGetTSbyVariable("Water Surface Elevation",1) 
}

}
