% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerGetTSVariables}
\alias{VeneerGetTSVariables}
\title{Get a vector of the type of time series variables recorded}
\usage{
VeneerGetTSVariables(run = "latest", baseURL = "http://localhost:9876")
}
\arguments{
\item{run}{Which run to retrieve from. Defaults to the latest}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
a vector of variable types (e.g. Downstream flow, Downstream Flow Concentration, water surface elevation)
}
\description{
Get a vector of the type of time series variables recorded
}
\examples{
\dontrun{
 VeneerGetTSVariables()
}

}
