% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerlatestRunNumber}
\alias{VeneerlatestRunNumber}
\title{Get the number of the latest run}
\usage{
VeneerlatestRunNumber(baseURL = "http://localhost:9876")
}
\arguments{
\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
integer of the latest run number
}
\description{
Get the number of the latest run
}
\examples{
\dontrun{
VeneerlatestRunNumber()
}

}
