% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerRunSource}
\alias{VeneerRunSource}
\title{Run Source using Veneer}
\usage{
VeneerRunSource(
  StartDate = NULL,
  EndDate = NULL,
  InputSet = NULL,
  baseURL = "http://localhost:9876"
)
}
\arguments{
\item{StartDate}{Optional. Start date for simulation. Must be dd/mm/yyyy}

\item{EndDate}{Optional. End date for simulation. Must be dd/mm/yyyy}

\item{InputSet}{Optional. Input set to use}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
Nothing to the R environment.

If not set, the configuration parameters (StartDate, EndDate, InputSet), was is specified
in the Source configuration in the GUI will be used.

The console will show any errors returned by Veneer.
}
\description{
Run Source using Veneer
}
\examples{
\dontrun{
VeneerRunSource()
VeneerRunSource("01/07/2017","01/02/2018","NoDams")
}

}
