% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerSetFunction}
\alias{VeneerSetFunction}
\title{Update a function value or expression. Function must exist before being updated.}
\usage{
VeneerSetFunction(Name, Expression, baseURL = "http://localhost:9876")
}
\arguments{
\item{Name}{Name of the function without the $, e.g. f_ScaleFactor}

\item{Expression}{Expression to change it to, e.g. 1.2}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\description{
Update a function value or expression. Function must exist before being updated.
}
\examples{
\dontrun{
 VeneerSetFunction("f_ScaleFactor",1.2)
 VeneerSetFunction("f_TargetLevel","if($m_Flow<1000,3.2,3.5)")
 }
 
}
