% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veneer.R
\name{VeneerSetPiecewise}
\alias{VeneerSetPiecewise}
\title{Change a Source piecewise table using Veneer}
\usage{
VeneerSetPiecewise(data, pw_table, baseURL = "http://localhost:9876")
}
\arguments{
\item{data}{A 2 column data.frame or matrix with the data to load into the piecewise table.}

\item{pw_table}{The name of the piecewise linear variable, without the "$".}

\item{baseURL}{URL of the Veneer server. Defaults to the veneer default.}
}
\value{
Nothing to the R environment.
}
\description{
Change a Source piecewise table using Veneer
}
\examples{
\dontrun{
 data<-data.frame(X=seq(1,5),Y=seq(1,5))
 VeneerSetPiecewise(data,"pw_table")
}
 
}
