% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{SILOLoad}
\alias{SILOLoad}
\title{Import multiple SILO files}
\usage{
SILOLoad(sites, path = getwd(), startdate, enddate)
}
\arguments{
\item{sites}{a vector of Station numbers (e.g. c("24001","24002","24003") to import. The function expects the file to be called "24001.txt".}

\item{path}{Location where the file is located. Use "/" or "\\\\" for folders. Defaults to getwd() if not specified.}

\item{startdate}{Start date of data to load, in format "YYYY-MM-DD". Defaults to start of the file if not provided}

\item{enddate}{End date of data to load, in format "YYYY-MM-DD". Defaults to end of the file if not provided}
}
\value{
a list of lists of SILO data from each file. Each list has members: tsd - the raw data as a daily zoo object, Site- the name of the site, Station - the station number, Lon- Longitude, and Lat - Latitude
}
\description{
Import multiple SILO files
}
\examples{
X<-LoadSILO(c("24001","24002","24003")
plot(X$tsd$Rain)
}
