% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO.R
\name{SILOReport}
\alias{SILOReport}
\title{Write SILO data report to word document}
\usage{
SILOReport(SILO, filename, path = getwd(), cols = pkg.env$col)
}
\arguments{
\item{SILO}{a list of sites with SILO data, as created by SILOLoad()}

\item{filename}{filename to write the report to.}

\item{path}{Optional. Folder to save the report to, defaults to current working directory}
}
\description{
Write SILO data report to word document
}
\examples{
X<-SILOLoad(c("24001","24002","24003"))
SILOReport(X,"C:/Output/MyReport.docx")

\\code{\\link{SILOLoad}}
}
