% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILOWriteforSource.R
\name{SILOWriteforSource}
\alias{SILOWriteforSource}
\title{Write a csv file that can be loaded into a model, e.g. Source}
\usage{
SILOWriteforSource(SILO, col, filename)
}
\arguments{
\item{SILO}{a list of sites with SILO data, as created by SILOLoad()}

\item{col}{Name of a column in a silo file to write out}

\item{filename}{file to write to.}
}
\description{
Write a csv file that can be loaded into a model, e.g. Source
}
\examples{
X<-LoadSILO(c("24001","24002","24003")
p<-SILOWriteforSource(X,"Rain","Rainfall.csv")
}
