% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILOWriteforSource.R
\name{WritepwtoIS}
\alias{WritepwtoIS}
\title{Write an input set line for a piecewise lookup table}
\usage{
WritepwtoIS(folder, csvfiles, outputfile)
}
\arguments{
\item{folder}{Folder for where are the csv files with the lookup table}

\item{csvfiles}{vector of files to turn into an input set line. 
File name should be the name of the pw table in Source, including the folder name if necessary. 
The first row should be column names  the same as used in Source, i.e. XValue and YValue}

\item{outputfile}{text file to save the lines to}
}
\description{
Write an input set line for a piecewise lookup table
}
\examples{
\dontrun{
folder<-"C:/Source/tables"
csvfiles<-c("LowerLakesOps.pw_LakeTarget.csv","Operations.pw_NA_Lock5_16p8.csv")
outputfile<-"inputset.txt"
WritepwtoIS(folder,csvfiles,outputfile)
}

}
